// contentful/generated/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful'
import { Document } from '@contentful/rich-text-types'

export interface ICategoryFields {
  /** 名称 */
  name: string
}

export interface ICategory extends Entry<ICategoryFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'category'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IContentFields {
  /** タイトル */
  title: string

  /** ハンズオン */
  isHandson: boolean

  /** カテゴリ */
  category: ICategory

  /** ターゲット */
  targets?: string[] | undefined

  /** オーナー */
  owner?: string | undefined

  /** 説明文 */
  description?: string | undefined

  /** プロダクト */
  products?: IProduct[] | undefined

  /** 言語 */
  languages?: ILanguage[] | undefined

  /** URL */
  url: string
}

export interface IContent extends Entry<IContentFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'content'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface ILanguageFields {
  /** 名称 */
  name: string
}

export interface ILanguage extends Entry<ILanguageFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'language'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IProductFields {
  /** 名称 */
  name: string
}

export interface IProduct extends Entry<IProductFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'product'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export type CONTENT_TYPE = 'category' | 'content' | 'language' | 'product'

export type LOCALE_CODE = 'en-US'

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US'